move confluence classes into confluence package
