+++ core/src/main/groovy/org/docToolchain/atlassian/confluence/ConfluenceService.groovy
@@ -1,4 +1,4 @@
-package org.docToolchain.atlassian
+package org.docToolchain.atlassian.confluence
 
 import org.docToolchain.configuration.ConfigService
 import org.jsoup.Jsoup
@@ -30,9 +30,9 @@ class ConfluenceService {
     Document parseFile(File htmlFile) {
         String html = htmlFile.getText('utf-8')
         Document dom = Jsoup.parse(html, 'utf-8', Parser.xmlParser())
-        dom.outputSettings().prettyPrint(false);//makes html() preserve linebreaks and spacing
-        dom.outputSettings().escapeMode(Entities.EscapeMode.xhtml); //This will ensure xhtml validity regarding entities
-        dom.outputSettings().charset("UTF-8"); //does no harm :-)
+        dom.outputSettings().prettyPrint(false)//makes html() preserve linebreaks and spacing
+        dom.outputSettings().escapeMode(Entities.EscapeMode.xhtml) //This will ensure xhtml validity regarding entities
+        dom.outputSettings().charset("UTF-8") //does no harm :-)
         return dom
     }
